Extract LinkedAttributeRenderer from AttributeRenderer
